// packages/web/src/views/Week/interaction/v2/model/AllDayResizeVisual.ts
import { type VisualPoint, type VisualRect } from "./TimedDragVisual";

export type AllDayResizeEdge = "startDate" | "endDate";

export interface AllDayResizeVisual {
  activeEdge: AllDayResizeEdge;
  endDayIndex: number;
  eventId: string;
  initialEndDayIndex: number;
  initialStartDayIndex: number;
  pointerStart: VisualPoint;
  sourceRect: VisualRect;
  startDayIndex: number;
  transform: VisualPoint;
  type: "allDayResize";
  width: number;
}